Extend Backwards-Compatible ClusterConfig to SSL
We're translating pre-akka-2.2-style transport-free remote actor paths
into 2.2-friendly paths. This extends that to SSL paths if
SiriusConfiguration.ENABLE_SSL is turned on, as well.

If the user enters akka-2.2-style paths, though, this does NOT change
those based on the ENABLE_SSL flag. That is, if they enter
'akka.ssl.tcp' and the ENABLE_SSL flag is off, this doesn't fix that for
them. I imagine that's just classified as a configuration error.