cleanup scaladoc
fixed various issues in the scaladoc to remove all but two of the
scaladoc warnings. The last two unresolved warnings relate to
linking to external scaladoc / javadoc. I could not figure out how to
get scaladoc to link to external documents without using absolute
urls. Will make this a separate issue in case we want to pursue this.