Disable quarantining, dial back health sensitivity
Quaranting actually seems to be detrimental to us right now. We maintain
our own cluster membership and test for connectivity, so the double-duty
is confusing matters. We're also seeing a bit too much churn with the
health checks, so dialing that back to a more generous level.

These could reasonably go back to defaults (or at least, closer to them)
if we changed the way we handle actor references (using ActorSelections,
perhaps) then we may want to adjust these again.