Catchup succeeds despite large compaction holes
Previously, compaction would stall and eventually fail if sufficiently
large compaction holes existed after and including the requested
sequence number. Specifically, if there were no updates within
MAX_WINDOW_SIZE of the requested sequence number, catchup would not be
able to make progress. This is a small change that will update nextSeq
properly in the face of empty catchup Subranges, allowing a follower to
recognize compaction holes and move on with its life.

Also added a couple of a FullSystemITests that demonstrate this problem.